[platform_device_loader] isolate recovery related operations
