+++ libparsec/crates/platform_device_loader/src/device/mod.rs
@@ -1,5 +1,7 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
 
 mod archive;
+mod recovery;
 
 pub use archive::*;
+pub use recovery::*;
